+++ node/CHANGELOG.md
@@ -11,6 +11,14 @@ All notable changes to this project will be documented in this file.  The format
 
 
 
+## 1.5.6
+
+### Changed
+* The node will recognise if a pending upgrade is unstaged and will avoid shutting down for upgrade in this case.
+* If an upgrade with the same activation point as the current one is detected on startup, the node will immediately shut down for upgrade.
+
+
+
 ## 1.5.5
 
 ### Added